+++ ios/batm/Entities/SellDetails.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct SellDetails: Equatable {
-  var dailyLimit: Double
-  var transactionLimit: Double
-  var profitRate: Double
+  var dailyLimit: Decimal
+  var transactionLimit: Decimal
+  var profitRate: Decimal
 }
